feat: Renaming metric chart, nav and stat. Adding top-nav component to index.
