+++ client/src/ui/metric-nav.tsx
